+++ src/IB/IBHierarchyIntegrator.C
@@ -3535,7 +3535,7 @@ IBHierarchyIntegrator::getFromInput(
         }
         else if (db->isDouble("min_ghost_cell_width"))
         {
-            d_ghosts = std::ceil(db->getDouble("min_ghost_cell_width"));
+            d_ghosts = int(std::ceil(db->getDouble("min_ghost_cell_width")));
         }
 
         d_start_time = db->getDoubleWithDefault("start_time", d_start_time);
