+++ src/IB/IBStaggeredHierarchyIntegrator.C
@@ -2629,7 +2629,7 @@ IBStaggeredHierarchyIntegrator::getFromInput(
         }
         else if (db->isDouble("min_ghost_cell_width"))
         {
-            d_ghosts = std::ceil(db->getDouble("min_ghost_cell_width"));
+            d_ghosts = int(std::ceil(db->getDouble("min_ghost_cell_width")));
         }
         d_using_pIB_method = db->getBoolWithDefault("using_pIB_method", d_using_pIB_method);
         if (d_using_pIB_method)
